Refactor to use filterable kwarg on QColormapComboBox
