+++ src/superqt/__init__.py
@@ -33,7 +33,6 @@ __all__ = [
     "QCollapsible",
     "QColorComboBox",
     "QColormapComboBox",
-    "QColormapFilterComboBox",
     "QDoubleRangeSlider",
     "QDoubleSlider",
     "QElidingLabel",
@@ -67,10 +66,6 @@ def __getattr__(name: str) -> Any:
         from .cmap import QColormapComboBox
 
         return QColormapComboBox
-    if name == "QColormapFilterComboBox":
-        from .cmap import QColormapFilterComboBox
-
-        return QColormapFilterComboBox
     if name == "QIconifyIcon":
         from .iconify import QIconifyIcon
