+++ src/superqt/cmap/__init__.py
@@ -10,7 +10,6 @@ else:
 
 from ._catalog_combo import CmapCatalogComboBox
 from ._cmap_combo import QColormapComboBox
-from ._cmap_filter_combo import QColormapFilterComboBox
 from ._cmap_item_delegate import QColormapItemDelegate
 from ._cmap_line_edit import QColormapLineEdit
 from ._cmap_utils import draw_colormap
@@ -18,7 +17,6 @@ from ._cmap_utils import draw_colormap
 __all__ = [
     "CmapCatalogComboBox",
     "QColormapComboBox",
-    "QColormapFilterComboBox",
     "QColormapItemDelegate",
     "QColormapLineEdit",
     "draw_colormap",
